Some commands into gef.commands package moved (aurusov/raox#334)
