+++ ru.bmstu.rk9.rao.ui/src/ru/bmstu/rk9/rao/ui/gef/commands/ChangeConstraintCommand.java
@@ -1,8 +1,8 @@
-package ru.bmstu.rk9.rao.ui.process.command;
+package ru.bmstu.rk9.rao.ui.gef.commands;
 
 import org.eclipse.draw2d.geometry.Rectangle;
 
-import ru.bmstu.rk9.rao.ui.process.node.Node;
+import ru.bmstu.rk9.rao.ui.gef.Node;
 
 public class ChangeConstraintCommand extends ConstraintCommand {
